+++ tests/__init__.py
@@ -1,6 +1,6 @@
 from .test_create import *
 from .test_parse import *
-from .test_encoding_error import *
+from .test_decoding_error import *
 from .test_encode import *
 from .test_decode import *
 from .test_hash_field import *